Move send_tx commands
