+++ zeroc/src/transaction/constants.rs
@@ -1,0 +1,2 @@
+pub const VERIFICATION_KEY_PATH: &str = "zeroc/verification.params";
+pub const PROVING_KEY_PATH: &str = "zeroc/proving.params";
