+++ zeroc/src/utils/mod.rs
@@ -1,3 +1,3 @@
-mod print_keys;
+pub mod print_keys;
 pub mod mnemonics;
 pub use self::print_keys::*;
